Rework analysis module (internal API and views blueprints)
