+++ bemserver_ui/templates/pages/analysis/energy_consumption/explore.html
@@ -82,7 +82,7 @@
 {{ super() -}}
 {% filter indent(width=8, first=True) %}
 <script type="module">
-    import { EnergyConsumptionExploreView } from "{{ url_for('static', filename='scripts/modules/views/analysis/energyConsumptionExplore.js') }}";
+    import { EnergyConsumptionExploreView } from "{{ url_for('static', filename='scripts/modules/views/analysis/energyConsumption/explore.js') }}";
 
 
     document.addEventListener("DOMContentLoaded", () => {
