+++ bemserver_ui/templates/pages/analysis/energy_consumption/setup.html
@@ -2,7 +2,7 @@
 {% extends "pages/base.html" %}
 
 {% set title = "Energy consumption" %}
-{% set subtitle = "Analysis configuration" %}
+{% set subtitle = "Analysis setup" %}
 
 {% set is_editable = signed_user.is_admin %}
 
@@ -99,12 +99,12 @@
 {% filter indent(width=8, first=True) %}
 <script type="module">
     import "{{ url_for('static', filename='scripts/modules/components/itemsCount.js') }}";
-    import { EnergyConsumptionConfigView } from "{{ url_for('static', filename='scripts/modules/views/analysis/energyConsumptionConfig.js') }}";
+    import { EnergyConsumptionSetupView } from "{{ url_for('static', filename='scripts/modules/views/analysis/energyConsumption/setup.js') }}";
 
 
     document.addEventListener("DOMContentLoaded", () => {
 
-        let energyConsConfigView = new EnergyConsumptionConfigView(
+        let view = new EnergyConsumptionSetupView(
             {{ structural_element | tojson }},
             {{ ener_cons_config | tojson }},
             {{ all_energy_sources | tojson }},
@@ -112,7 +112,7 @@
             {{ available_energy_sources | tojson }},
             {{ is_editable | tojson }},
         );
-        energyConsConfigView.refresh();
+        view.refresh();
 
     });
 </script>
